+++ database/migrations/2018_11_26_155324_create_habitacions_table.php
@@ -15,12 +15,12 @@ class CreateHabitacionsTable extends Migration
     {
         Schema::create('habitaciones', function (Blueprint $table) {
             $table->increments('id');
-            $table->unsignedBigInteger('hotel_id');
+            $table->integer('hotel_id');
             $table->foreign('hotel_id')->references('id')->on('hoteles');
             $table->smallInteger('numero_habitacion');
             $table->smallInteger('capacidad');
             $table->text('descripcion');
-            $table->unsignedSmallInteger('precio');
+            $table->unsignedInteger('precio');
             $table->timestamps();
         });
     }
